requirements.txt
asgiref==3.11.0
dj-database-url==2.1.0
Django==4.2.15
django-filter==23.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
sqlparse==0.5.5
